+++ RELEASE_NOTES.md
@@ -9,6 +9,9 @@
 - Added callback for random color button.
 - Fixed callbacks so that they work appropriately via an event system. (#88)
 - Changed mutability of widget_store, get_widget_for_id now returns a RefCell so widgets can be accessed directly.
+- Added ability to look up a widget by name. (#94)
+- Added ability to add a widget, and a widget with parent with an assigned name.
+- Internally used "_WidgetStoreBase" as the internal first, or base widget.  Updated docs to reflect this.
 
 ## 0.2.3